Implemented "fit to page" zoom

+++ src/gui/ZoomHandler.hh
@@ -17,14 +17,14 @@ class ZoomHandler : public Glib::Object
 public:
   ZoomHandler(JournalWidget& parent_);
 
-  Glib::PropertyProxy_ReadOnly<float> property_zoom_level() const
+  Glib::PropertyProxy_ReadOnly<double> property_zoom_level() const
   {
-    return Glib::PropertyProxy_ReadOnly<float>(this, "zoom-level");
+    return Glib::PropertyProxy_ReadOnly<double>(this, "zoom-level");
   }
 
-  float get_zoom_level() const;
+  double get_zoom_level() const;
 
-  void set_zoom_level(float value);
+  void set_zoom_level(double value);
 
   void zoom_in();
   void zoom_out();
@@ -35,13 +35,13 @@ public:
 private:
   void on_layout_size_allocated(Gtk::Allocation& alloc);
 
-  Glib::Property<float> prop_zoom_level;
+  Glib::Property<double> prop_zoom_level;
   JournalWidget& parent;
 
   sigc::connection layout_con;
 
   LayoutPosition layout_pos;
-  float fx, fy;
+  double fx, fy;
 };
 
 #endif
